requirements.txt
absl-py==0.10.0
astor==0.8.1
beautifulsoup4==4.6.0
bs4==0.0.1
certifi==2020.6.20
chardet==3.0.4
fonttools==4.14.0
gast==0.4.0
google-pasta==0.2.0
grpcio==1.32.0
h5py==2.10.0
idna==2.10
importlib-metadata==1.7.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
lxml==4.5.2
Markdown==3.2.2
numpy==1.16.0
opencv-python==4.4.0.42
Pillow==7.2.0
protobuf==3.13.0
psutil==5.7.2
requests==2.24.0
selenium==3.141.0
six==1.15.0
soupsieve==2.0.1
tb-nightly==1.14.0a20190603
tensorflow==2.0.0b0
termcolor==1.1.0
tf-estimator-nightly==1.14.0.dev2019060501
urllib3==1.25.10
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.1.0
